+++ R/tidyverse.R
@@ -128,28 +128,35 @@ use_tidy_description <- function() {
 #' @rdname tidyverse
 #' @param overwrite By default (`FALSE`), only dependencies without version
 #'   specifications will be modified. Set to `TRUE` to modify all dependencies.
-use_tidy_versions <- function(overwrite = FALSE) {
+#' @param source Use "local" or "CRAN" package versions.
+use_tidy_versions <- function(overwrite = FALSE, source = c("local", "CRAN")) {
   deps <- desc::desc_get_deps(proj_get())
+  deps <- update_versions(deps, overwrite = overwrite, source = source)
+  desc::desc_set_deps(deps, file = proj_get())
 
+  invisible(TRUE)
+}
+
+update_versions <- function(deps, overwrite = FALSE, source = c("local", "CRAN")) {
   baserec <- base_and_recommended()
   to_change <- !deps$package %in% c("R", baserec)
   if (!overwrite) {
     to_change <- to_change & deps$version == "*"
   }
 
-  deps$version[to_change] <- vapply(deps$package[to_change], dep_version, character(1))
-  desc::desc_set_deps(deps, file = proj_get())
+  packages <- deps$package[to_change]
+  versions <- switch(match.arg(source),
+    local = purrr::map_chr(packages, package_version),
+    CRAN = utils::available.packages()[packages, "Version"]
+  )
+  deps$version[to_change] <- paste0(">= ", versions)
 
-  invisible(TRUE)
+  deps
 }
 
-is_installed <- function(x) {
-  length(find.package(x, quiet = TRUE)) > 0
+package_version <- function(x) {
+  as.character(utils::packageVersion(x))
 }
-dep_version <- function(x) {
-  if (is_installed(x)) paste0(">= ", utils::packageVersion(x)) else "*"
-}
-
 
 #' @export
 #' @rdname tidyverse